+++ src/strategy/mod.rs
@@ -18,7 +18,7 @@
 //!
 //! * [`DefaultStrategy`] (this one is used implicitly)
 //! * [`IndependentStrategy`]
-//! * [`RwLock<()>][std::sync::RwLock]
+//! * [`RwLock<()>`][std::sync::RwLock]
 //!
 //! # Testing
 //!
@@ -34,6 +34,13 @@
 //!
 //! *This is not meant to be used in production code*.
 //!
+//! # Experimental strategies
+//!
+//! There are some more strategies inside the [`experimental`] module. Note that these **are not**
+//! subject to the API stability guarantees and can be changed, renamed or removed at any time.
+//!
+//! They are available only with the `experimental-strategies` feature.
+//!
 //! [`ArcSwap`]: crate::ArcSwap
 //! [`load`]: crate::ArcSwapAny::load
 
@@ -43,9 +50,13 @@ use std::sync::atomic::AtomicPtr;
 use crate::gen_lock::{Global, PrivateUnsharded};
 use crate::ref_cnt::RefCnt;
 
+#[cfg(feature = "experimental-strategies")]
+pub mod experimental;
+mod gen_lock;
 mod hybrid;
 mod rw_lock;
 
+use self::gen_lock::GenLockStrategy;
 use self::hybrid::HybridStrategy;
 
 /// The default strategy.
@@ -74,7 +85,7 @@ use self::hybrid::HybridStrategy;
 /// similar or better properties.
 ///
 /// [`load`]: crate::ArcSwapAny::load
-pub type DefaultStrategy = HybridStrategy<Global>;
+pub type DefaultStrategy = HybridStrategy<GenLockStrategy<Global>>;
 
 /// Strategy for isolating instances.
 ///
@@ -90,7 +101,7 @@ pub type DefaultStrategy = HybridStrategy<Global>;
 ///
 /// This too may be changed for something else (but with at least as good guarantees, primarily
 /// that other instances won't get influenced by the "torture").
-pub type IndependentStrategy = HybridStrategy<PrivateUnsharded>;
+pub type IndependentStrategy = HybridStrategy<GenLockStrategy<PrivateUnsharded>>;
 
 // TODO: When we are ready to un-seal, should these traits become unsafe?